Removed umbraco-marketplace tag from core project
Thanks for marking up this package for inclusion in the Umbraco marketplace.  I noticed though we were bringing in two entries - on for the installable component of the package, and one for this "core" project, which likely shouldn't be listed.  So I've suggested this update so we can just include the main package that developers will usually choose to install.